+++ crates/zng-color/src/gradient.rs
@@ -1062,9 +1062,11 @@ impl GradientStops {
         // length that must be split between positional stops.
         let (start_offset, layout_length) = {
             // index of stop after the sequence that has a calculated offset.
-            let sequence_ender = (range.end..render_stops.len()).find(|i| !hints.contains(i)).unwrap();
+            let sequence_ender = (range.end..render_stops.len())
+                .find(|i| !hints.contains(i))
+                .unwrap_or(range.end - 1);
             // index of stop before the sequence that has a calculated offset.
-            let sequence_starter = (0..range.start).rev().find(|i| !hints.contains(i)).unwrap();
+            let sequence_starter = (0..range.start).rev().find(|i| !hints.contains(i)).unwrap_or(range.start);
 
             let start_offset = render_stops[sequence_starter].offset;
             let length = render_stops[sequence_ender].offset - start_offset;
@@ -1075,7 +1077,7 @@ impl GradientStops {
         let mut offset = start_offset;
 
         for i in range {
-            if !hints.contains(&i) {
+            if ColorStop::is_layout_positional(render_stops[i].offset) {
                 offset += d;
                 render_stops[i].offset = offset;
             }